Refactor user store path and implement user state management
- Updated import paths for useUserStore in useAuth and AuthProvider components.
- Introduced userStore to manage user state, including fetching user data and handling authentication status.
